get stepper to update value

+++ LyricFinder/LyricFinder/Model/BHSong.h
@@ -12,11 +12,11 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface BHSong : NSObject
 
-@property NSString* UUID;
+@property NSString *UUID;
 @property NSString *title;
 @property NSString *artist;
 @property NSString *lyrics;
-@property int rating;
+@property double rating;
 
 @end
